+++ lib/fetch.bash
@@ -96,7 +96,7 @@ function fetch_pip_requirements {
     # -e ./lib/python-project
     #
     local req_dir=$(dirname $pip_requirements);
-    (cd $req_dir && execu pip install -r $pip_requirements);
+    (cd $req_dir && exec_pip install -r $pip_requirements);
 }
 
 # fetch_oca_requirements <filepath>
@@ -170,7 +170,7 @@ function fetch_python_dep {
         local install_opt="$1";
     fi
 
-    execu pip install $install_opt;
+    exec_pip install $install_opt;
 }