+++ lib/server.bash
@@ -26,6 +26,7 @@ function get_server_script {
     check_command odoo odoo.py openerp-server openerp-server.py;
 }
 
+
 # Function to check server run status;
 # Function echo:
 #   pid - server running process <pid>